wifi: rtw89: hw_scan: Don't let the operating channel be last
Scanning can be offloaded to the firmware. To that end, the driver
prepares a list of channels to scan, including periodic visits to
the operating channel, and sends the list to the firmware.

When the channel list is too long to fit in a single H2C message,
the driver splits the list, sends the first part, and tells the
firmware to scan. When the scan is complete, the driver sends the
next part of the list and tells the firmware to scan.

When the last channel that fit in the H2C message is the operating
channel something seems to go wrong in the firmware. It will
acknowledge receiving the list of channels but apparently it will
not do anything more. The AP can't be pinged anymore. The driver
still receives beacons, though.

One way to avoid this is to split the list of channels before the
operating channel.

Another way is to comment out "ch_info->tx_null = true;" but then
the pings are higher during scans.

Affected devices:

* RTL8851BU with firmware 0.29.41.3
* RTL8832BU and RTL8852BE with firmware 0.29.29.8

Not affected:

* RTL8832BU and RTL8852BE with firmware 1.29.29.9

Well, RTL8852BE behaves a bit differently: the driver detects a
firmware crash (?) and tries to restart the firmware.

Signed-off-by: Bitterblue Smith <rtl8821cerfe2@gmail.com>